restructure linker and fix memory leaks
