+++ compiler/linker/linker.c
@@ -4,139 +4,159 @@
 
 #include "util.h"
 #include "symbols.h"
+#include "input.h"
 
-// returns 0 if export, 1 if require
-char parseLinkDirection(char *directionString)
+void linkerParseFile(FILE *inFile, char *inFileName, struct Dictionary *symbolNames, struct LinkedList *exports[s_null], struct LinkedList *requires[s_null])
 {
-    if (!strcmp(directionString, "require"))
-    {
-        return 1;
-    }
-    else if (!strcmp(directionString, "export"))
-    {
-        return 0;
-    }
-    else
-    {
-        ErrorAndExit(ERROR_INTERNAL, "Linker Error - unexpected direction %s\n", directionString);
-    }
-}
+    int nSymbols[s_null];
+    memset(nSymbols, 0, s_null * sizeof(int));
 
-void addRequire(struct LinkedList **exports, struct LinkedList **requires, struct Symbol *toRequire)
-{
-    enum LinkedSymbol symbolType = toRequire->symbolType;
-    if (LinkedList_Find(requires[symbolType], compareSymbols, toRequire) == NULL)
-    {
-        LinkedList_Append(requires[symbolType], toRequire);
-    }
-    else
-    {
-        Symbol_Free(toRequire);
-    }
-}
+    int len;
+    char requireNewSymbol = 1;
 
-void addExport(struct LinkedList **exports, struct LinkedList **requires, struct Symbol *toAdd)
-{
-    enum LinkedSymbol addType = toAdd->symbolType;
+    char currentLinkDirection = -1;
+    enum LinkedSymbol currentLinkSymbolType = s_null;
+    struct Symbol *currentSymbol = NULL;
 
-    // check for duplicates of anything except function declarations
-    struct Symbol *found;
-    if ((addType != s_function_declaration) &&
-        (found = LinkedList_Find(exports[addType], compareSymbols, toAdd)))
+    char ignoreDuplicateSymbol = 0;
+    while (!feof(inFile))
     {
-        // if we see something other than section userstart duplicated, throw an error
-        if ((strcmp(toAdd->name, "userstart") || (addType != s_section)))
+        len = getline_force_raw(&inBuf, &bufSize, inFile);
+        if (len == -1)
         {
-
-            ErrorAndExit(ERROR_CODE, "Multiple definition of symbol %s %s - from %s and %s!\n", symbolEnumToName(addType), toAdd->name, found->fromFile, toAdd->fromFile);
+            break;
         }
 
-        LinkedList_Join(found->lines, toAdd->lines);
+        if (inBuf[0] == '~')
+        {
+            // Extract the first token
+            char *token = strtok(inBuf + 1, " ");
+            if (!strcmp(token, "end"))
+            {
+                if (requireNewSymbol)
+                {
+                    ErrorAndExit(ERROR_INTERNAL, "Linker error - expected new symbol to start but got end instead!\n");
+                }
 
-        // printf("\n\n\n\ndelete duplicate start lines\n");
-        // while (LinkedList_Find(found->lines, strcmp, "START:"))
+                token = strtok(NULL, " ");
+                if (parseLinkDirection(token) != currentLinkDirection)
+                {
+                    ErrorAndExit(ERROR_INTERNAL, "Unexpected end directive - link direction doesn't match start!\n");
+                }
+
+                token = strtok(NULL, " ");
+                if (symbolNameToEnum(token) != currentLinkSymbolType)
+                {
+                    ErrorAndExit(ERROR_INTERNAL, "End symbol type doesn't match start!\n");
+                }
+
+                token = strtok(NULL, " ");
+                if (strcmp(token, currentSymbol->name))
+                {
+                    ErrorAndExit(ERROR_INTERNAL, "End symbol name (%s) doesn't match start (%s)!\n", token, currentSymbol->name);
+                }
+
+                if(ignoreDuplicateSymbol)
+                {
+                    Symbol_Free(currentSymbol);
+                }
+
+                requireNewSymbol = 1;
+            }
+            else
+            {
+                if (!requireNewSymbol)
+                {
+                    ErrorAndExit(ERROR_INTERNAL, "Started reading new symbol when not expected!\n");
+                }
+                requireNewSymbol = 0;
+
+                currentLinkDirection = parseLinkDirection(token);
+
+                token = strtok(NULL, " ");
+                currentLinkSymbolType = symbolNameToEnum(token);
+
+                token = strtok(NULL, " ");
+
+                currentSymbol = Symbol_New(Dictionary_LookupOrInsert(symbolNames, token), currentLinkDirection, currentLinkSymbolType, inFileName);
+
+                switch (currentLinkSymbolType)
+                {
+                case s_function_declaration:
+                    // fall through to function definition as the symbol data is identical for declaration and definition
+                case s_function_definition:
+                {
+                    printf("function dec/def\n");
+                    parseFunctionDeclaration(currentSymbol, inFile);
+                }
+                break;
+
+                case s_variable:
+                {
+                    printf("variable\n");
+                    parseVariable(currentSymbol, inFile);
+                }
+                break;
+
+                case s_section:
+                    break;
+
+                case s_object:
+                {
+                    printf("object\n");
+                    parseObject(currentSymbol, inFile);
+                }
+                break;
+
+                case s_null:
+                    ErrorAndExit(ERROR_INTERNAL, "Saw s_null as link symbol type!\n");
+                    break;
+                }
+
+                if (currentLinkDirection == export)
+                {
+                    ignoreDuplicateSymbol = addExport(exports, requires, currentSymbol);
+                }
+                else
+                {
+                    addRequire(exports, requires, currentSymbol);
+                    ignoreDuplicateSymbol = 0;
+                }
+                nSymbols[currentLinkSymbolType]++;
+            }
+        }
+        else // copying data, just stick the line onto the existing WIP symbol
+        {
+            if (currentSymbol == NULL)
+            {
+                ErrorAndExit(ERROR_INVOCATION, "Malformed input file - couldn't find directive!\n");
+            }
+            
+            LinkedList_Append(currentSymbol->lines, strTrim(inBuf, len));
+        }
+        // loop through the string to extract all other tokens
+        // while (token != NULL)
         // {
-        // LinkedList_Delete(found->lines, strcmp, "START:");
+        // printf(" %s\n", token); // printing each token
+        // token = strtok(NULL, " ");
         // }
-
-        // Symbol_Free(toAdd);
-        return;
+        // printf("\n");
     }
 
-    LinkedList_Append(exports[addType], toAdd);
-
-    // if we export a declaration, we must require a definition for it as well
-    if (addType == s_function_declaration)
+    printf("\t");
+    for (int i = 0; i < s_null; i++)
     {
-        struct Symbol *funcDefRequired = Symbol_New(toAdd->name, require, s_function_definition, toAdd->fromFile);
-        funcDefRequired->data.asFunction = toAdd->data.asFunction;
-
-        funcDefRequired->data.asFunction.args = malloc(toAdd->data.asFunction.nArgs * sizeof(struct Type));
-        memcpy(funcDefRequired->data.asFunction.args, toAdd->data.asFunction.args, toAdd->data.asFunction.nArgs * sizeof(struct Type));
-
-        funcDefRequired->data.asFunction.returnType = malloc(sizeof(struct Type));
-        memcpy(funcDefRequired->data.asFunction.returnType, toAdd->data.asFunction.returnType, sizeof(struct Type));
-
-        for (struct LinkedListNode *runner = toAdd->lines->head; runner != NULL; runner = runner->next)
-        {
-            LinkedList_Append(funcDefRequired->lines, strdup(runner->data));
-        }
-
-        for (struct LinkedListNode *runner = toAdd->linkerLines->head; runner != NULL; runner = runner->next)
-        {
-            LinkedList_Append(funcDefRequired->linkerLines, strdup(runner->data));
-        }
-
-        addRequire(exports, requires, funcDefRequired);
+        printf("%d %s ", nSymbols[i], symbolEnumToName(i));
     }
-
-    // if adding this export satisfies any requires, delete them
-    if (LinkedList_Find(requires[addType], compareSymbols, toAdd) != NULL)
-    {
-        struct Symbol *deletedRequire = LinkedList_Delete(requires[addType], compareSymbols, toAdd);
-        Symbol_Free(deletedRequire);
-    }
-}
-
-int getline_force_raw(char **linep, size_t *linecapp, FILE *stream)
-{
-    int len = getline(linep, linecapp, stream);
-    if (len == 0)
-    {
-        ErrorAndExit(ERROR_INTERNAL, "getline_force expects non-zero line length, got 0!\n");
-    }
-    if (len == -1)
-    {
-        return len;
-    }
-
-    if ((*linep)[len - 1] == '\n')
-    {
-        (*linep)[len - 1] = '\0';
-        len--;
-    }
-    return len;
-}
-
-size_t getline_force(char **linep, size_t *linecapp, FILE *stream, struct Symbol *currentSymbol)
-{
-    size_t len = getline_force_raw(linep, linecapp, stream);
-    LinkedList_Append(currentSymbol->lines, strTrim(*linep, len));
-    return len;
-}
-
-size_t getline_force_metadata(char **linep, size_t *linecapp, FILE *stream, struct Symbol *currentSymbol)
-{
-    size_t len = getline_force_raw(linep, linecapp, stream);
-    LinkedList_Append(currentSymbol->linkerLines, strTrim(*linep, len));
-    return len;
+    printf("\n");
 }
 
 int main(int argc, char **argv)
 {
     struct LinkedList *exports[s_null];
     struct LinkedList *
-        requires[s_null];
+        requires[
+            s_null];
 
     for (int i = 0; i < s_null; i++)
     {
@@ -185,8 +205,20 @@ int main(int argc, char **argv)
 
     printf("have %d input files\n", inputFiles->buckets[0]->size);
 
-    char *inBuf = NULL;
-    size_t bufSize = 0;
+    if (outputExecutable)
+    {
+        struct Symbol *main = Symbol_New(Dictionary_LookupOrInsert(symbolNames, "main"), require, s_function_definition, "");
+        main->data.asFunction.nArgs = 0;
+
+        struct Type *mainReturnType = malloc(sizeof(struct Type));
+        mainReturnType->isPrimitive = 1;
+        mainReturnType->size = 0;
+        mainReturnType->indirectionLevel = 0;
+
+        main->data.asFunction.returnType = mainReturnType;
+
+        addRequire(exports, requires, main);
+    }
 
     for (struct LinkedListNode *inFileName = inputFiles->buckets[0]->head; inFileName != NULL; inFileName = inFileName->next)
     {
@@ -197,301 +229,7 @@ int main(int argc, char **argv)
             ErrorAndExit(ERROR_INTERNAL, "Error opening file %s\n", (char *)inFileName->data);
         }
 
-        int nSymbols[s_null];
-        memset(nSymbols, 0, s_null * sizeof(int));
-
-        if (outputExecutable)
-        {
-            struct Symbol *main = Symbol_New(Dictionary_LookupOrInsert(symbolNames, "main"), require, s_function_definition, inFileName->data);
-            main->data.asFunction.nArgs = 0;
-
-            struct Type *mainReturnType = malloc(sizeof(struct Type));
-            mainReturnType->isPrimitive = 1;
-            mainReturnType->size = 0;
-            mainReturnType->indirectionLevel = 0;
-
-            main->data.asFunction.returnType = mainReturnType;
-
-            addRequire(exports, requires, main);
-        }
-
-        int len;
-        char requireNewSymbol = 1;
-
-        char currentLinkDirection = -1;
-        enum LinkedSymbol currentLinkSymbolType = s_null;
-        struct Symbol *currentSymbol = NULL;
-
-        while (!feof(inFile))
-        {
-            len = getline_force_raw(&inBuf, &bufSize, inFile);
-            if (len == -1)
-            {
-                break;
-            }
-
-            if (inBuf[0] == '~')
-            {
-                // Extract the first token
-                char *token = strtok(inBuf + 1, " ");
-                if (!strcmp(token, "end"))
-                {
-                    if (requireNewSymbol)
-                    {
-                        ErrorAndExit(ERROR_INTERNAL, "Linker error - expected new symbol to start but got end instead!\n");
-                    }
-
-                    token = strtok(NULL, " ");
-                    if (parseLinkDirection(token) != currentLinkDirection)
-                    {
-                        ErrorAndExit(ERROR_INTERNAL, "Unexpected end directive - link direction doesn't match start!\n");
-                    }
-
-                    token = strtok(NULL, " ");
-                    if (symbolNameToEnum(token) != currentLinkSymbolType)
-                    {
-                        ErrorAndExit(ERROR_INTERNAL, "End symbol type doesn't match start!\n");
-                    }
-
-                    token = strtok(NULL, " ");
-                    if (strcmp(token, currentSymbol->name))
-                    {
-                        ErrorAndExit(ERROR_INTERNAL, "End symbol name (%s) doesn't match start (%s)!\n", token, currentSymbol->name);
-                    }
-                    requireNewSymbol = 1;
-                }
-                else
-                {
-                    if (!requireNewSymbol)
-                    {
-                        ErrorAndExit(ERROR_INTERNAL, "Started reading new symbol when not expected!\n");
-                    }
-                    requireNewSymbol = 0;
-
-                    currentLinkDirection = parseLinkDirection(token);
-
-                    token = strtok(NULL, " ");
-                    currentLinkSymbolType = symbolNameToEnum(token);
-
-                    token = strtok(NULL, " ");
-
-                    currentSymbol = Symbol_New(Dictionary_LookupOrInsert(symbolNames, token), currentLinkDirection, currentLinkSymbolType, inFileName->data);
-
-                    switch (currentLinkSymbolType)
-                    {
-                    case s_function_declaration:
-                        // ErrorAndExit(ERROR_INTERNAL, "Function declaration not yet supported!\n");
-                        // break;
-                        // fall through to function definition as the symbol data is identical for declaration and definition
-
-                    case s_function_definition:
-                    {
-                        len = getline_force_metadata(&inBuf, &bufSize, inFile, currentSymbol);
-                        char *token = strtok(inBuf, " ");
-                        if (strcmp(token, "returns"))
-                        {
-                            ErrorAndExit(ERROR_INTERNAL, "Expected returns [type] but got %s instead!\n", token);
-                        }
-
-                        currentSymbol->data.asFunction.returnType = parseType(inBuf + strlen(token) + 1);
-
-                        len = getline_force_metadata(&inBuf, &bufSize, inFile, currentSymbol);
-                        token = strtok(inBuf, " ");
-
-                        int nArgs = atoi(token);
-                        currentSymbol->data.asFunction.nArgs = nArgs;
-
-                        if (nArgs > 0)
-                        {
-                            currentSymbol->data.asFunction.args = malloc(nArgs * sizeof(struct Type));
-                        }
-
-                        token = strtok(NULL, " ");
-                        if (strcmp(token, "arguments"))
-                        {
-                            ErrorAndExit(ERROR_INTERNAL, "Expected '[n] arguments' but got %s %s instead!\n", inBuf, token);
-                        }
-
-                        while (nArgs-- > 0)
-                        {
-                            len = getline_force_metadata(&inBuf, &bufSize, inFile, currentSymbol);
-                        }
-                    }
-                    break;
-
-                    case s_variable:
-                    {
-                        getline_force_metadata(&inBuf, &bufSize, inFile, currentSymbol);
-                        struct Type *varType = parseType(inBuf);
-                        currentSymbol->data.asVariable = *varType;
-                        free(varType);
-                    }
-                    break;
-
-                    case s_section:
-                        break;
-
-                    case s_object:
-                    {
-                        getline_force_metadata(&inBuf, &bufSize, inFile, currentSymbol);
-
-                        char *token = strtok(inBuf, " ");
-                        if (strcmp(token, "size"))
-                        {
-                            ErrorAndExit(ERROR_INTERNAL, "Expected size [size] but got %s instead!\n", token);
-                        }
-                        token = strtok(NULL, " ");
-
-                        currentSymbol->data.asObject.size = atoi(token);
-
-                        token = strtok(NULL, " ");
-                        if (strcmp(token, "initialized"))
-                        {
-                            ErrorAndExit(ERROR_INTERNAL, "Expected initialized [1/0] but got %s instead!\n", token);
-                        }
-                        token = strtok(NULL, " ");
-                        currentSymbol->data.asObject.isInitialized = (atoi(token) == 1);
-
-                        if (currentSymbol->data.asObject.isInitialized)
-                        {
-                            size_t nRead = getline_force_metadata(&inBuf, &bufSize, inFile, currentSymbol);
-
-                            if (currentSymbol->data.asObject.size * 3 != nRead)
-                            {
-                                ErrorAndExit(ERROR_INTERNAL, "Bad formatting of byte stream for object initialization!\n");
-                            }
-                            currentSymbol->data.asObject.initializeTo = malloc(currentSymbol->data.asObject.size);
-                            size_t charIndex = 0;
-                            size_t byteIndex = 0;
-                            while (byteIndex < currentSymbol->data.asObject.size)
-                            {
-                                unsigned char byte = 0;
-                                for (int i = 0; i < 2; i++)
-                                {
-                                    switch (inBuf[charIndex])
-                                    {
-                                    case '0':
-                                        byte |= 0;
-                                        break;
-
-                                    case '1':
-                                        byte |= 1;
-                                        break;
-
-                                    case '2':
-                                        byte |= 2;
-                                        break;
-
-                                    case '3':
-                                        byte |= 3;
-                                        break;
-
-                                    case '4':
-                                        byte |= 4;
-                                        break;
-
-                                    case '5':
-                                        byte |= 5;
-                                        break;
-
-                                    case '6':
-                                        byte |= 6;
-                                        break;
-
-                                    case '7':
-                                        byte |= 7;
-                                        break;
-
-                                    case '8':
-                                        byte |= 8;
-                                        break;
-
-                                    case '9':
-                                        byte |= 9;
-                                        break;
-
-                                    case 'a':
-                                        byte |= 10;
-                                        break;
-
-                                    case 'b':
-                                        byte |= 11;
-                                        break;
-
-                                    case 'c':
-                                        byte |= 12;
-                                        break;
-
-                                    case 'd':
-                                        byte |= 13;
-                                        break;
-
-                                    case 'e':
-                                        byte |= 14;
-                                        break;
-
-                                    case 'f':
-                                        byte |= 15;
-                                        break;
-
-                                    default:
-                                        ErrorAndExit(ERROR_INTERNAL, "Malformed hex input for object initialization data!\nGot char '%c', expected valid hex char\n", inBuf[charIndex]);
-                                    }
-
-                                    charIndex++;
-                                    // only shift on the first char
-                                    byte <<= (1 - i);
-                                }
-                                if (inBuf[charIndex] != ' ')
-                                {
-                                    ErrorAndExit(ERROR_INTERNAL, "Malformed hex input for object initialization data!\nSaw '%c', expected ' '\n", inBuf[charIndex]);
-                                }
-                                charIndex++;
-                                currentSymbol->data.asObject.initializeTo[byteIndex++] = byte;
-                            }
-                        }
-                    }
-                    break;
-
-                    case s_null:
-                        ErrorAndExit(ERROR_INTERNAL, "Saw s_null as link symbol type!\n");
-                        break;
-                    }
-
-                    if (currentLinkDirection == export)
-                    {
-                        addExport(exports, requires, currentSymbol);
-                    }
-                    else
-                    {
-                        addRequire(exports, requires, currentSymbol);
-                    }
-                    nSymbols[currentLinkSymbolType]++;
-                }
-            }
-            else // copying data, just stick the line onto the existing WIP symbol
-            {
-                if (currentSymbol == NULL)
-                {
-                    ErrorAndExit(ERROR_INVOCATION, "Malformed input file - couldn't find directive!\n");
-                }
-                LinkedList_Append(currentSymbol->lines, strTrim(inBuf, len));
-            }
-            // loop through the string to extract all other tokens
-            // while (token != NULL)
-            // {
-            // printf(" %s\n", token); // printing each token
-            // token = strtok(NULL, " ");
-            // }
-            // printf("\n");
-        }
-
-        printf("\t");
-        for (int i = 0; i < s_null; i++)
-        {
-            printf("%d %s ", nSymbols[i], symbolEnumToName(i));
-        }
-        printf("\n");
+        linkerParseFile(inFile, (char *)inFileName->data, symbolNames, exports, requires);
     }
 
     free(inBuf);
@@ -508,9 +246,9 @@ int main(int argc, char **argv)
     int nOutputRequirements = 0;
     for (int i = 0; i < s_null; i++)
     {
-        if (requires[i] -> size > 0)
+        if (requires[i]->size > 0)
         {
-            for (struct LinkedListNode *runner = requires[i] -> head; runner != NULL; runner = runner->next)
+            for (struct LinkedListNode *runner = requires[i]->head; runner != NULL; runner = runner->next)
             {
                 nOutputRequirements++;
             }
@@ -524,9 +262,9 @@ int main(int argc, char **argv)
             printf("Missing definitions:\n");
             for (int i = 0; i < s_null; i++)
             {
-                if (requires[i] -> size > 0)
+                if (requires[i]->size > 0)
                 {
-                    for (struct LinkedListNode *runner = requires[i] -> head; runner != NULL; runner = runner->next)
+                    for (struct LinkedListNode *runner = requires[i]->head; runner != NULL; runner = runner->next)
                     {
                         struct Symbol *required = runner->data;
                         printf("\t%s %s\n", symbolEnumToName(required->symbolType), required->name);
@@ -540,9 +278,9 @@ int main(int argc, char **argv)
 
         for (int i = 0; i < s_null; i++)
         {
-            if (requires[i] -> size > 0)
+            if (requires[i]->size > 0)
             {
-                for (struct LinkedListNode *runner = requires[i] -> head; runner != NULL; runner = runner->next)
+                for (struct LinkedListNode *runner = requires[i]->head; runner != NULL; runner = runner->next)
                 {
                     struct Symbol *required = runner->data;
                     fprintf(outFile, "~require %s %s\n", symbolEnumToName(required->symbolType), required->name);
@@ -560,7 +298,7 @@ int main(int argc, char **argv)
 
     for (int i = 0; i < s_null; i++)
     {
-        // printf("%d %s(s)\n", exports[i]->size, symbolEnumToName(i));
+        printf("%d %s(s)\n", exports[i]->size, symbolEnumToName(i));
         if (exports[i]->size > 0)
         {
             for (struct LinkedListNode *runner = exports[i]->head; runner != NULL; runner = runner->next)
@@ -572,6 +310,7 @@ int main(int argc, char **argv)
                 }
 
                 Symbol_Write(exported, outFile, outputExecutable);
+                printf("\t%s\n", exported->name);
                 if (!outputExecutable)
                 {
                     fprintf(outFile, "~end export %s %s\n", symbolEnumToName(exported->symbolType), exported->name);
